requirements.txt
spotipy==2.23.0
telethon==1.34.0
requests==2.31.0
python-dotenv==1.0.0
colorama==0.4.6
tqdm==4.66.1
cryptg==0.4.0